Split generate command
