+++ cmd/util-db/generate/db_hash.go
@@ -1,0 +1,48 @@
+package generate
+
+import (
+	"fmt"
+	"github.com/0xsoniclabs/aida/logger"
+	"github.com/0xsoniclabs/aida/utildb"
+	"github.com/0xsoniclabs/aida/utils"
+	"github.com/0xsoniclabs/substate/db"
+	"github.com/urfave/cli/v2"
+)
+
+var DbHashCommand = cli.Command{
+	Action: generateDbHashAction,
+	Name:   "generate-db-hash",
+	Usage:  "Generates new db-hash. Note that this will overwrite the current AidaDb hash.",
+	Flags: []cli.Flag{
+		&utils.AidaDbFlag,
+	},
+}
+
+// generateDbHashAction calculates the dbHash for given AidaDb and saves it.
+func generateDbHashAction(ctx *cli.Context) error {
+	log := logger.NewLogger("INFO", "DbHashGenerateCMD")
+
+	cfg, err := utils.NewConfig(ctx, utils.NoArgs)
+
+	aidaDb, err := db.NewDefaultBaseDB(cfg.AidaDb)
+	if err != nil {
+		return fmt.Errorf("cannot open db; %v", err)
+	}
+
+	defer utildb.MustCloseDB(aidaDb)
+
+	md := utils.NewAidaDbMetadata(aidaDb, "INFO")
+
+	log.Noticef("Starting DbHash generation for %v; this may take several hours...", cfg.AidaDb)
+	hash, err := utildb.GenerateDbHash(aidaDb, "INFO")
+	if err != nil {
+		return err
+	}
+
+	err = md.SetDbHash(hash)
+	if err != nil {
+		return fmt.Errorf("cannot set db-hash; %v", err)
+	}
+
+	return nil
+}
